feat: add error handling

+++ src/main/java/com/mathdev/quickbite/services/OrderService.java
@@ -11,6 +11,11 @@ import com.mathdev.quickbite.entities.User;
 import com.mathdev.quickbite.mapper.OrderMapper;
 import com.mathdev.quickbite.repositories.OrderRepository;
 import com.mathdev.quickbite.repositories.UserRepository;
+import com.mathdev.quickbite.services.exceptions.DatabaseException;
+import com.mathdev.quickbite.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
+
 import com.mathdev.quickbite.entities.Order;
 
 @Service
@@ -24,41 +29,69 @@ public class OrderService {
 
 	// GET SERVICES
 	public List<OrderDTO> findAll() {
-		List<Order> orders = repo.findAll();
+		try {
+			List<Order> orders = repo.findAll();
 
-		return orders.stream().map(OrderMapper::toDTO).collect(Collectors.toList()); // fitful for Java < 16 versions
+			return orders.stream().map(OrderMapper::toDTO).collect(Collectors.toList()); // fitful for Java < 16
+																							// versions
+		} catch (Exception e) {
+			throw new DatabaseException("Error retrieving products: " + e.getMessage());
+		}
 	}
 
 	public OrderDTO findById(Long id) {
-		Order obj = repo.findById(id).orElseThrow(() -> new RuntimeException("User not found!"));
+		Order obj = repo.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Order with id" + id + " not found!"));
 
 		return OrderMapper.toDTO(obj);
 	}
 
 	public List<OrderDTO> findOrdersByUserId(Long userId) {
-		List<Order> list = repo.findByClientId(userId);
-		return list.stream().map(OrderMapper::toDTO).collect(Collectors.toList());
+		try {
+			List<Order> list = repo.findByClientId(userId);
+			return list.stream().map(OrderMapper::toDTO).collect(Collectors.toList());
+		} catch (Exception e) {
+			throw new DatabaseException("Error retrieving orders by user: " + e.getMessage());
+		}
 	}
 
 	// POST SERVICES
 	public OrderDTO insert(OrderDTO dto) {
-		Order obj = OrderMapper.fromDTO(dto);
-		obj = repo.save(obj);
+		try {
+			Order obj = OrderMapper.fromDTO(dto);
 
-		return OrderMapper.toDTO(obj);
+			obj = repo.save(obj);
+
+			return OrderMapper.toDTO(obj);
+		} catch (Exception e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 
 	// DELETE SERVICES
-	public void deleteUser(Long id) {
-		repo.deleteById(id);
+	public void deleteOrder(Long id) {
+		if (!repo.existsById(id)) {
+			throw new ResourceNotFoundException("Order id " + id + " does not exist");
+		}
+		try {
+			repo.deleteById(id);
+
+		} catch (Exception e) {
+			throw new DatabaseException("Error deleting order: " + e.getMessage());
+		}
 	}
 
 	// PUT SERVICES
 	public OrderDTO update(Long id, OrderDTO newData) {
-		Order entity = repo.getReferenceById(id);
-		updateData(entity, newData);
+		try {
+			Order entity = repo.getReferenceById(id);
+			updateData(entity, newData);
 
-		return OrderMapper.toDTO(repo.save(entity));
+			return OrderMapper.toDTO(repo.save(entity));
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
+
 	}
 
 	// AUXILIARY METHODS
@@ -72,4 +105,3 @@ public class OrderService {
 	}
 
 }
-
